requirements_live.txt
# Live Trading System Requirements
# Install with: pip install -r requirements_live.txt

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data APIs
fredapi>=0.5.0
yfinance>=0.2.0

# Broker integrations (install only what you need)
# oandapyV20>=0.7.0        # For OANDA
# ib_insync>=0.9.86        # For Interactive Brokers
# alpaca-trade-api>=3.0.0  # For Alpaca

# Utilities
schedule>=1.2.0  # For scheduled tasks
pytz>=2023.3     # Timezone handling

# Monitoring & Alerts
# twilio>=8.0.0           # For SMS alerts (optional)
# slack-sdk>=3.0.0        # For Slack notifications (optional)

# Visualization (for dashboard)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Optional: Advanced features
# scikit-learn>=1.3.0     # If using ML models
# scipy>=1.11.0           # For optimization
